+++ client/reviews/index.jsx
@@ -1,0 +1,20 @@
+import React from 'react';
+import ReactDom from 'react-dom';
+
+class Reviews extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {};
+  }
+
+  render() {
+    return (
+      <div>
+        Reviews
+      </div>
+    )
+  }
+}
+
+
+export default Reviews;
