+++ package.json
@@ -2,20 +2,25 @@
   "scripts": {
     "react-dev": "webpack --mode development --watch",
     "server-dev": "nodemon server/index.js",
+    "test": "jest",
     "start": "node server/index.js"
   },
   "dependencies": {
     "axios": "^0.24.0",
     "bluebird": "^3.7.2",
     "body-parser": "^1.19.0",
+    "chai": "^4.3.4",
     "cors": "^2.8.5",
+    "css-loader": "^6.5.1",
     "eslint": "^8.4.1",
     "express": "^4.17.1",
+    "mini-css-extract-plugin": "^2.4.5",
     "moment": "^2.29.1",
     "mongoose": "^6.1.2",
     "nodemon": "^2.0.15",
     "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react-dom": "^17.0.2",
+    "style-loader": "^3.3.1"
   },
   "devDependencies": {
     "@babel/cli": "^7.16.0",
@@ -23,6 +28,9 @@
     "@babel/preset-env": "^7.16.4",
     "@babel/preset-react": "^7.16.0",
     "babel-loader": "^8.2.3",
+    "enzyme": "^3.11.0",
+    "jest": "^27.4.5",
+    "mocha": "^9.1.3",
     "webpack": "^5.65.0",
     "webpack-cli": "^4.9.1"
   }
