+++ server/index.js
@@ -10,5 +10,7 @@ app.use(express.static(path.join(__dirname, "/client/main")));
 app.use(cors());
 app.use(bodyParser.json());
 
+app.get('/')
+
 // this is for a test
 app.listen(port);
