+++ webpack.config.js
@@ -1,4 +1,5 @@
 var path = require('path');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 
 module.exports = {
   entry: path.join(__dirname, "client/app.jsx"),
@@ -16,3 +17,31 @@ module.exports = {
     }]
   }
 }
+
+// module.exports = {
+//   plugins: [
+//     new MiniCssExtractPlugin({
+//       filename: 'stylesheet.css'
+//     })
+//   ],
+//   entry: path.join(__dirname, "client/app.jsx"),
+//   output: {
+//     filename: "bundle.js",
+//     path: path.join(__dirname, "client/main")
+//   },
+//   module: {
+//     rules: [{
+//       test: /\.jsx?/,
+//       exclude: /node_modules/,
+//       use: {
+//         loader: "babel-loader"
+//       }
+//     }, {
+//       test: /\.css$/,
+//       use: [
+//         'style-loader',
+//         'css-loader'
+//       ]
+//     }]
+//   }
+// }
